+++ tests-integration/tests/auction_app.rs
@@ -280,6 +280,7 @@ fn test_auction_app_recipient() {
         andr.kernel.addr(),
         None,
         None,
+        None,
     );
     let splitter_component = AppComponent::new(
         "splitter",
@@ -338,7 +339,7 @@ fn test_auction_app_recipient() {
         None,
         None,
         None,
-        Some(Recipient::from_string("./splitter").with_msg(mock_splitter_send_msg())),
+        Some(Recipient::from_string("./splitter").with_msg(mock_splitter_send_msg(None))),
     );
     cw721
         .execute_send_nft(