+++ tests-integration/tests/crowdfund_app.rs
@@ -92,8 +92,13 @@ fn setup(
             percent: Decimal::from_str("0.8").unwrap(),
         },
     ];
-    let splitter_init_msg =
-        mock_splitter_instantiate_msg(splitter_recipients, andr.kernel.addr().clone(), None, None);
+    let splitter_init_msg = mock_splitter_instantiate_msg(
+        splitter_recipients,
+        andr.kernel.addr().clone(),
+        None,
+        None,
+        None,
+    );
     let splitter_component = AppComponent::new(
         "splitter".to_string(),
         "splitter".to_string(),
@@ -840,6 +845,6 @@ fn mock_recipient_with_invalid_msg(addr: &str) -> Recipient {
 fn mock_recipient_with_valid_msg(addr: &str) -> Recipient {
     Recipient::new(
         addr,
-        Some(to_json_binary(&mock_splitter_send_msg()).unwrap()),
+        Some(to_json_binary(&mock_splitter_send_msg(None)).unwrap()),
     )
 }